All errors solved
